+++ src/dflow/__init__.py
@@ -1,4 +1,4 @@
-from .workflow import Workflow
+from .workflow import Workflow, config
 from .step import Step, argo_range, argo_sequence, argo_len, if_expression
 from .io import (
     AutonamedDict,
@@ -12,7 +12,7 @@ from .io import (
 )
 from .steps import Steps
 from .op_template import OPTemplate, ShellOPTemplate, PythonScriptOPTemplate
-from .utils import upload_s3, download_s3, copy_s3, upload_artifact, download_artifact, copy_artifact
+from .utils import s3_config, upload_s3, download_s3, copy_s3, upload_artifact, download_artifact, copy_artifact
 from .common import S3Artifact
 from .executor import Executor, RemoteExecutor
 from .resource import Resource